Add Otter operations from BuildMaster to Otter.
